pt_region_create function moved to main ittapi module
