+++ python/ittapi/compat/__init__.py
@@ -4,5 +4,5 @@ Python binding to NERSC implementation of Intel Instrumentation and Tracing Tech
 from ittapi.native import pt_region_begin, pt_region_end
 from ittapi.collection_control import resume, pause, detach
 from ittapi.domain import domain as domain_create
+from ittapi.pt_region import pt_region_create
 from .task import task_begin, task_end
-from .pt_region import pt_region_create
